requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
bcrypt==4.1.2
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.6
distro==1.9.0
docarray==0.40.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==2.4
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.2
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.63
markdown-it-py==3.0.0
marshmallow==3.21.2
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.30.5
orjson==3.10.3
packaging==23.2
paramiko==3.4.0
pathspec==0.12.1
platformdirs==4.2.1
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-osc==1.8.3
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
rich==13.7.1
sniffio==1.3.1
SQLAlchemy==2.0.30
sshtunnel==0.4.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
types-requests==2.32.0.20240523
typing-inspect==0.9.0
typing_extensions==4.12.0
urllib3==2.2.1
yarl==1.9.4
